create Generator class
